Add curl example for Travelpayouts token
