+++ tests/test_server.py
@@ -1,8 +1,7 @@
 """Tests for the FastMCP server wrapper."""
-import sys
-import os
-
 from __future__ import annotations
+
+import sys
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))  # noqa: E402
